+++ CHANGELOG.md
@@ -9,6 +9,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 
 ### Added
 
+- `--role` flag to `teams invite`
 - `teams join` using an invitation token
 
 ### Fixed